requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.6.0
drf-yasg
